FIR/UAST: commonize USimpleReferenceExpression
as accessor call to synthetic Java property becomes frontend-agnostic
