+++ plugins/kotlin/uast/uast-kotlin-base/kotlin.uast.uast-kotlin-base.iml
@@ -30,6 +30,7 @@
     <orderEntry type="module" module-name="intellij.java.psi.impl" />
     <orderEntry type="module" module-name="intellij.java.rt" scope="TEST" />
     <orderEntry type="module" module-name="intellij.java.uast" />
+    <orderEntry type="module" module-name="kotlin.fir.frontend-independent" />
     <orderEntry type="library" scope="TEST" name="kotlin-test" level="project" />
     <orderEntry type="module" module-name="intellij.platform.uast.tests" scope="TEST" />
     <orderEntry type="module" module-name="kotlin.tests-common" scope="TEST" />
