--- plugins/kotlin/uast/uast-kotlin-base/src/org/jetbrains/uast/kotlin/expressions/KotlinAbstractUSimpleReferenceExpression.kt
@@ -1,40 +1,0 @@
-// Copyright 2000-2021 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
-
-package org.jetbrains.uast.kotlin
-
-import com.intellij.openapi.util.Key
-import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiNamedElement
-import org.jetbrains.kotlin.psi.KtSimpleNameExpression
-import org.jetbrains.kotlin.psi.UserDataProperty
-import org.jetbrains.uast.UElement
-import org.jetbrains.uast.USimpleNameReferenceExpression
-import org.jetbrains.uast.internal.acceptList
-import org.jetbrains.uast.toUElement
-import org.jetbrains.uast.visitor.UastVisitor
-
-var PsiElement.destructuringDeclarationInitializer: Boolean? by UserDataProperty(Key.create("kotlin.uast.destructuringDeclarationInitializer"))
-
-abstract class KotlinAbstractUSimpleReferenceExpression(
-    override val sourcePsi: KtSimpleNameExpression,
-    givenParent: UElement?
-) : KotlinAbstractUExpression(givenParent), USimpleNameReferenceExpression, KotlinUElementWithType, KotlinEvaluatableUElement {
-    protected val resolvedDeclaration: PsiElement? by lz { baseResolveProviderService.resolveToDeclaration(sourcePsi) }
-
-    override val identifier get() = sourcePsi.getReferencedName()
-
-    override fun resolve() = resolvedDeclaration
-
-    override val resolvedName: String?
-        get() = (resolvedDeclaration as? PsiNamedElement)?.name
-
-    override fun accept(visitor: UastVisitor) {
-        visitor.visitSimpleNameReferenceExpression(this)
-
-        uAnnotations.acceptList(visitor)
-
-        visitor.afterVisitSimpleNameReferenceExpression(this)
-    }
-
-    override val referenceNameElement: UElement? by lz { sourcePsi.getIdentifier()?.toUElement() }
-}
