--- plugins/kotlin/uast/uast-kotlin-fir/src/org/jetbrains/uast/kotlin/expressions/FirKotlinUSimpleReferenceExpression.kt
@@ -1,14 +1,0 @@
-// Copyright 2000-2021 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
-
-package org.jetbrains.uast.kotlin.expressions
-
-import org.jetbrains.kotlin.psi.KtSimpleNameExpression
-import org.jetbrains.uast.UElement
-import org.jetbrains.uast.kotlin.KotlinAbstractUSimpleReferenceExpression
-
-class FirKotlinUSimpleReferenceExpression(
-    override val sourcePsi: KtSimpleNameExpression,
-    givenParent: UElement?
-) : KotlinAbstractUSimpleReferenceExpression(sourcePsi, givenParent) {
-    // TODO: handle destructuring declaration differently or commonize it
-}
