+++ plugins/kotlin/uast/uast-kotlin-fir/testData/legacyRenderLog/Resolve.log.fir.txt
@@ -66,6 +66,9 @@ UFile (package = )
                     UQualifiedReferenceExpression
                         USimpleNameReferenceExpression (identifier = date)
                         USimpleNameReferenceExpression (identifier = time)
+                            UCallExpression (kind = UastCallKind(name='method_call'), argCount = 1))
+                                UIdentifier (Identifier (time))
+                                ULiteralExpression (value = 1000)
                     ULiteralExpression (value = 1000)
                 UQualifiedReferenceExpression
                     UCallExpression (kind = UastCallKind(name='method_call'), argCount = 1))
