+++ plugins/kotlin/uast/uast-kotlin-fir/testData/legacyValues/Resolve.values.fir.txt
@@ -75,6 +75,9 @@ UFile (package = ) [public final class ResolveKt {...]
                     UQualifiedReferenceExpression [date.time] = external time()
                         USimpleNameReferenceExpression (identifier = date) [date] = (var date = external <init>()())
                         USimpleNameReferenceExpression (identifier = time) [time] = external time()
+                            UCallExpression (kind = UastCallKind(name='method_call'), argCount = 1)) [setTime(1000)] = external setTime(1000)((long)1000)
+                                UIdentifier (Identifier (time)) [UIdentifier (Identifier (time))]
+                                ULiteralExpression (value = 1000) [1000] = (long)1000
                     ULiteralExpression (value = 1000) [1000] = (long)1000
                 UQualifiedReferenceExpression [listOf("").last()] = external last()()
                     UCallExpression (kind = UastCallKind(name='method_call'), argCount = 1)) [listOf("")] = external listOf("")("")
